module/common: sw_timer macro range extend & sw_timer LSE sleep time extend
Test Plan: No

Reviewers: jwzhang, mingzhou

Reviewed By: jwzhang

Differential Revision: http://10.10.20.8/D998

+++ dev/module/common/sw_timer.h
@@ -4,8 +4,8 @@
 #include <stdbool.h>
 #include "circular_doubly_linked_list.h"
 #ifdef LE501X
-#define MS_2_PERIOD(x) (2000*(x)/625)
-#define PERIOD_2_MS(x) (625*(x)/2000)
+#define MS_2_PERIOD(x) (2000*(uint64_t)(x)/625)
+#define PERIOD_2_MS(x) (625*(uint64_t)(x)/2000)
 #elif !defined(GEMINI)
 #include "ll_port.h"
 #define MS_2_PERIOD(x) ((x)*MAC_CLK_MHZ*1000)
